Parse new private package format
The new private package format is `{slug}@{username}`.
A version can also be appended (`{slug}@{username}@{version}`).

Normal packages work the same, including with versions. Note that
valid versions for packages either start with a non-letter character
or it is `be` or `stable` (which are accounted for as neither can be a
valid username).