Set the server name when creating new app (#307)
The name is not set by default in the server.json in the app templates, so it will use the current directory as the server name. That might be ok for some, but you can also end up with a bunch of servers named `app` or `wwwroot`, etc. if you create a sub dir in your project for the web root.