update description and links
The current description did not indicate CommandBox's role as offering an "embedded CFML server", which is the wording as indicated in the getting started guide linked to here. That link also was wailing, so I have corrected that. 

Also, I updated a link or two that redirected to what seemed clearly different URLs reached via redirects, so just go to the redirect. I did NOT change such redirect URLs that served as a front controller to some current, specific version (like the API docs link).